tostify added and delete added

+++ src/Pages/Todo/TodoList.js
@@ -8,20 +8,22 @@ const TodoList = ({ toDoLists, isLoading, refetch }) => {
     }
 
     const handleDelete = (id) => {
-        fetch(`http://localhost:8000/todolist/${id}`, {
-            method: 'DELETE',
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-                toast.success(`Data deleted`)
-                refetch();
+        const deleteData = window.confirm("Are you want delete the item?");
+        if (deleteData) {
+            fetch(`http://localhost:8000/todolist/${id}`, {
+                method: 'DELETE',
             })
+                .then(res => res.json())
+                .then(data => {
+                    console.log(data)
+                    toast.success(`Item Deleted Successfully`)
+                    refetch();
+                })
+        }
     }
 
     return (
-        <div className='w-2/4 mx-auto'>
-            <h2>{toDoLists.length}</h2>
+        <div className='w-2/4 mx-auto mt-4'>
             <div class="overflow-x-auto">
                 <table class="table w-full">
                     <thead>
@@ -40,7 +42,9 @@ const TodoList = ({ toDoLists, isLoading, refetch }) => {
                                     <th>{index + 1}</th>
                                     <td>{todo.todoName}</td>
                                     <td>{todo.toDoDescription}</td>
-                                    <td><button onClick={() => handleDelete(todo._id)} class="btn btn-sm btn-error text-white">Delete</button></td>
+                                    <td><button
+                                        onClick={() => handleDelete(todo._id)}
+                                        class="btn btn-sm btn-error text-white">Delete</button></td>
                                     <td><button class="btn btn-sm btn-accent text-white">Complete</button></td>
                                 </tr>)
                         }
